Renaming utils folder to shared
