+++ src/components/GCCMarks.vue
@@ -30,10 +30,10 @@
 <script>
 import { ref } from "vue";
 import { defineComponent } from "vue";
-import BrowserStorage from '@/utils/storage'
-import { GCCMarkEmojiMap } from '@/utils/emojis/GCCMarkEmojiMap'
-import AnalyticsManager from "@/utils/analytics";
-import { dynamicEvents } from "@/utils/analytics/events";
+import BrowserStorage from '@/shared/storage'
+import { GCCMarkEmojiMap } from '@/shared/emojis/GCCMarkEmojiMap'
+import AnalyticsManager from "@/shared/analytics";
+import { dynamicEvents } from "@/shared/analytics/events";
 
 const availableMarks = {
   'friendly': false,
