+++ src/scripts/lobby/LobbiesModifier.ts
@@ -1,18 +1,18 @@
 // @ts-check
-import { cleanSelector } from '@/utils/StringUtils'
+import { cleanSelector } from '@/shared/StringUtils'
 import $ from 'jquery'
 import { domEntityType } from './domain/domEntityType'
 import { createApp } from 'vue'
 import KDRComponent from '../../components/KDR.vue'
 import GCCMarkComponent from '../../components/GCCMarks.vue'
 import GCCLobbyTitle from '../../components/GCCLobbyTitle.vue'
-import { gcSelectors } from '../../utils/gc/selectors'
+import { gcSelectors } from '../../shared/gc/selectors'
 import serializer from './serializer'
 import lobbyFilter from './lobbyFilter'
 import Logger from 'js-logger'
-import BrowserStorage from '../../utils/storage'
-import gcBooster from '@/utils/gcBooster'
-import { newTippy } from '@/utils/tippy'
+import BrowserStorage from '../../shared/storage'
+import gcBooster from '@/shared/gcBooster'
+import { newTippy } from '@/shared/tippy'
 import Lobby from './domain/Lobby'
 import GCCLobbyPinner from '../../components/GCCLobbyPinner.vue'
 import LobbyPlayer from './domain/LobbyPlayer'
