+++ src/shared/StringUtils.ts
