+++ src/shared/analytics/events.ts
