+++ src/shared/devAuxScripts/README.md
