+++ src/shared/devAuxScripts/checkGCSelectors.js
