+++ src/shared/emojis/GCCMarkEmojiMap.ts
