+++ src/shared/gc/api/resources/constants/magicNumbers.ts
