+++ src/shared/gc/api/resources/urls/index.ts
