+++ src/shared/gc/types/GCInitialPlayerStats.ts
