+++ src/shared/storage/types.ts
